Rethink of 2014 and 2015
The first thinking on Dec 19 was more like the list of the github projects
that I care about. I will rewrite that post in Chinese for the projects:
what are they and why I care about them.

Two days ago, I realized it was better to summarize my year 2014 from
both life and work for the important things; and make new year's resolution
a balance of life and work too. So I will rethink of these in Chinese.