Plan to check automation rate and applicable to Crosswalk
As required we need to know the automation rate of these tests for
Chromium for Android and whether applicable to Crosswalk.

Will try to compare these tests with Crosswalk Projects integration
to see whether they have been integrated; if yes, applicable to
Crosswalk and then check the automation rate and testing time consume;
otherwise, not applicable. This is a thought stroke me while sleeping
with my son yesterday night.