Plan to write down my exploring Battery Status API
My initial thought on 20160124 was to write an article of using
Battery Status API; but it seems a little duplicate to existing ones.

Yesterday when I tried my sample to monitor battery status to debug
a test file in web-platform-tests repository, I realized it may be
valuable to write down my experience of the test suite and inter-
operability testing.