Try to share what learned from a book in 10 minutes
The idea of reading a book in 10 minutes (4000 words) comes from
xiaolai-xuexi, a wechat public account.

JavaScript: The Definitive Guide, is a great book for refreshing
my knowledge of core JavaScript and client-side JavaScript.

Software Testing Foundations, is a good book for refreshing testing
skills, which is also required to train new hires.