Get a Design & Systems Engineering Q1 2016 Department Award
for Crosswalk Project for Linux Deepin Release

Deepin is the leading desktop OS vendor in PRC, with many significant
customers in government and enterprise. This team’s success at
making Crosswalk the default web runtime in the Deepin 2015 formal
release was a major step to extend Crosswalk’s influence into the top
desktop OS and in promoting the HTML5 ecosystem to desktop developers.